Enable score categories
Use score categories in tags instead of real scores

+++ src/tag.rs
@@ -7,7 +7,7 @@ use serde::{Deserialize, Serialize};
 pub struct Tag {
     pub commitment: Vec<u8>,
     pub exp_timestamp: i64,
-    pub score: i32,
+    pub score: i8,
     pub enc_sender_id: Vec<u8>,
     pub q_big: RistrettoPoint,
     pub g_prime: RistrettoPoint,
